python_vuln_app/requirements.txt
PyYAML==5.3
email==3.11.3